Project struct changes
