+++ IOS4_HW13_MIKHAIL_BAGMET/Sources/Scenes/Albums/Controllers/CollectionViewCells/PlacesCollectionViewCell.swift
